Initial Commit for streamlit addon
Plan is to reuse the JS interface being used in toga webview as the streamlit interface as well.
Having gotten back data, we need to update the backend to handle the base64 image directly, instead of using file based approach.
Also we need to do something about the widget height settings